refactor: refactor the stateful tx validator to get executable tx as input
