+++ crates/gateway/src/stateful_transaction_validator_test.rs
@@ -5,11 +5,8 @@ use blockifier::blockifier::stateful_validator::{
 use blockifier::context::BlockContext;
 use blockifier::test_utils::CairoVersion;
 use blockifier::transaction::errors::{TransactionFeeError, TransactionPreValidationError};
-use mempool_test_utils::invoke_tx_args;
 use mempool_test_utils::starknet_api_test_utils::{
-    deploy_account_tx,
-    invoke_tx,
-    rpc_invoke_tx,
+    executable_invoke_tx,
     TEST_SENDER_ADDRESS,
     VALID_L1_GAS_MAX_AMOUNT,
     VALID_L1_GAS_MAX_PRICE_PER_UNIT,
@@ -19,7 +16,7 @@ use num_bigint::BigUint;
 use pretty_assertions::assert_eq;
 use rstest::{fixture, rstest};
 use starknet_api::core::{ContractAddress, Nonce, PatriciaKey};
-use starknet_api::rpc_transaction::RpcTransaction;
+use starknet_api::executable_transaction::Transaction;
 use starknet_api::transaction::{Resource, TransactionHash};
 use starknet_api::{contract_address, felt, patricia_key};
 use starknet_gateway_types::errors::GatewaySpecError;
@@ -66,18 +63,16 @@ fn stateful_validator(block_context: BlockContext) -> StatefulTransactionValidat
 // TODO(Arni): consider testing declare and deploy account.
 #[rstest]
 #[case::valid_tx(
-    invoke_tx(CairoVersion::Cairo1),
+    executable_invoke_tx(CairoVersion::Cairo1),
     Ok(ValidateInfo{
-        tx_hash: TransactionHash(felt!(
-        "0x3c6546364f70395ad7c8d1ce64b990b698be21e786bee2efff57a694f37648e"
-        )),
+        tx_hash: TransactionHash::default(),
         sender_address: contract_address!("0xc0020000"),
         account_nonce: Nonce::default()
     })
 )]
-#[case::invalid_tx(invoke_tx(CairoVersion::Cairo1), Err(STATEFUL_VALIDATOR_FEE_ERROR))]
+#[case::invalid_tx(executable_invoke_tx(CairoVersion::Cairo1), Err(STATEFUL_VALIDATOR_FEE_ERROR))]
 fn test_stateful_tx_validator(
-    #[case] rpc_tx: RpcTransaction,
+    #[case] executable_tx: Transaction,
     #[case] expected_result: BlockifierStatefulValidatorResult<ValidateInfo>,
     stateful_validator: StatefulTransactionValidator,
 ) {
@@ -90,7 +85,7 @@ fn test_stateful_tx_validator(
     mock_validator.expect_validate().return_once(|_, _| expected_result.map(|_| ()));
     mock_validator.expect_get_nonce().returning(|_| Ok(Nonce(Felt::ZERO)));
 
-    let result = stateful_validator.run_validate(&rpc_tx, None, mock_validator);
+    let result = stateful_validator.run_validate(&executable_tx, mock_validator);
     assert_eq!(result, expected_result_as_stateful_transaction_result);
 }
 
@@ -130,28 +125,45 @@ fn test_instantiate_validator() {
 
 #[rstest]
 #[case::should_skip_validation(
-    rpc_invoke_tx(invoke_tx_args!{nonce: Nonce(Felt::ONE)}),
+    Transaction::Invoke(starknet_api::test_utils::invoke::executable_invoke_tx(
+        starknet_api::invoke_tx_args!(nonce: Nonce(Felt::ONE))
+    )),
     Nonce::default(),
     true
 )]
 #[case::should_not_skip_validation_nonce_over_max_nonce_for_skip(
-    rpc_invoke_tx(invoke_tx_args!{nonce: Nonce(Felt::TWO)}),
+    Transaction::Invoke(starknet_api::test_utils::invoke::executable_invoke_tx(
+        starknet_api::invoke_tx_args!(nonce: Nonce(Felt::ZERO))
+    )),
     Nonce::default(),
     false
 )]
-#[case::should_not_skip_validation_non_invoke(deploy_account_tx(), Nonce::default(), false)]
+#[case::should_not_skip_validation_non_invoke(
+    Transaction::DeployAccount(
+        starknet_api::test_utils::deploy_account::executable_deploy_account_tx(
+            starknet_api::deploy_account_tx_args!(), Nonce::default()
+        )
+    ),
+    Nonce::default(),
+    false)]
 #[case::should_not_skip_validation_account_nonce_1(
-    rpc_invoke_tx(invoke_tx_args!{sender_address: ContractAddress::from(TEST_SENDER_ADDRESS), nonce: Nonce(Felt::ONE)}),
+    Transaction::Invoke(starknet_api::test_utils::invoke::executable_invoke_tx(
+        starknet_api::invoke_tx_args!(
+            nonce: Nonce(Felt::ONE),
+            sender_address: TEST_SENDER_ADDRESS.into()
+        )
+    )),
     Nonce(Felt::ONE),
     false
 )]
 fn test_skip_stateful_validation(
-    #[case] rpc_tx: RpcTransaction,
+    #[case] executable_tx: Transaction,
     #[case] sender_nonce: Nonce,
     #[case] should_skip_validate: bool,
     stateful_validator: StatefulTransactionValidator,
 ) {
-    let sender_address = rpc_tx.calculate_sender_address().unwrap();
+    let sender_address = executable_tx.contract_address();
+
     let mut mock_validator = MockStatefulTransactionValidatorTrait::new();
     mock_validator
         .expect_get_nonce()
@@ -161,5 +173,5 @@ fn test_skip_stateful_validation(
         .expect_validate()
         .withf(move |_, skip_validate| *skip_validate == should_skip_validate)
         .returning(|_, _| Ok(()));
-    let _ = stateful_validator.run_validate(&rpc_tx, None, mock_validator);
+    let _ = stateful_validator.run_validate(&executable_tx, mock_validator);
 }
