+++ crates/gateway/src/utils.rs
@@ -1,37 +1,18 @@
-use blockifier::execution::contract_class::ClassInfo;
 use blockifier::transaction::account_transaction::AccountTransaction;
-use blockifier::transaction::transactions::{
-    DeclareTransaction as BlockifierDeclareTransaction,
-    DeployAccountTransaction as BlockifierDeployAccountTransaction,
-    InvokeTransaction as BlockifierInvokeTransaction,
-};
-use starknet_api::core::{calculate_contract_address, ChainId, ClassHash, ContractAddress};
+use starknet_api::core::{ChainId, ContractAddress};
 use starknet_api::executable_transaction::{
     DeclareTransaction as ExecutableDeclareTransaction,
     DeployAccountTransaction as ExecutableDeployAccountTransaction,
     InvokeTransaction as ExecutableInvokeTransaction,
     Transaction as ExecutableTransaction,
 };
-use starknet_api::rpc_transaction::{
-    RpcDeclareTransaction,
-    RpcDeployAccountTransaction,
-    RpcInvokeTransaction,
-    RpcTransaction,
-};
-use starknet_api::transaction::{
-    DeclareTransaction,
-    DeclareTransactionV3,
-    DeployAccountTransaction,
-    DeployAccountTransactionV3,
-    InvokeTransactionV3,
-    TransactionHasher,
-    ValidResourceBounds,
-};
+use starknet_api::rpc_transaction::{RpcDeclareTransaction, RpcTransaction};
+use starknet_api::transaction::DeclareTransaction;
 use starknet_gateway_types::errors::GatewaySpecError;
 use tracing::{debug, error};
 
 use crate::compilation::GatewayCompiler;
-use crate::errors::{GatewayResult, StatefulTransactionValidatorResult};
+use crate::errors::GatewayResult;
 
 /// Converts an RPC transaction to an executable transaction.
 /// Note, for declare transaction this step is heavy, as it requires compilation of Sierra to
@@ -96,105 +77,6 @@ fn compile_contract_and_build_executable_declare_tx(
         })?;
 
     Ok(executable_declare_tx)
-}
-
-// TODO(Arni): Remove this function.
-pub fn rpc_tx_to_account_tx(
-    rpc_tx: &RpcTransaction,
-    // FIXME(yael 15/4/24): calculate class_info inside the function once compilation code is ready
-    optional_class_info: Option<ClassInfo>,
-    chain_id: &ChainId,
-) -> StatefulTransactionValidatorResult<AccountTransaction> {
-    match rpc_tx {
-        RpcTransaction::Declare(RpcDeclareTransaction::V3(tx)) => {
-            let declare_tx = DeclareTransaction::V3(DeclareTransactionV3 {
-                class_hash: ClassHash::default(), /* FIXME(yael 15/4/24): call the starknet-api
-                                                   * function once ready */
-                resource_bounds: ValidResourceBounds::AllResources(tx.resource_bounds),
-                tip: tx.tip,
-                signature: tx.signature.clone(),
-                nonce: tx.nonce,
-                compiled_class_hash: tx.compiled_class_hash,
-                sender_address: tx.sender_address,
-                nonce_data_availability_mode: tx.nonce_data_availability_mode,
-                fee_data_availability_mode: tx.fee_data_availability_mode,
-                paymaster_data: tx.paymaster_data.clone(),
-                account_deployment_data: tx.account_deployment_data.clone(),
-            });
-            let tx_hash = declare_tx
-                .calculate_transaction_hash(chain_id, &declare_tx.version())
-                .map_err(|e| {
-                    error!("Failed to calculate tx hash: {}", e);
-                    GatewaySpecError::UnexpectedError { data: "Internal server error".to_owned() }
-                })?;
-            let class_info =
-                optional_class_info.expect("declare transaction should contain class info");
-            let declare_tx = BlockifierDeclareTransaction::new(declare_tx, tx_hash, class_info)
-                .map_err(|e| {
-                    error!("Failed to convert declare tx hash to blockifier tx type: {}", e);
-                    GatewaySpecError::UnexpectedError { data: "Internal server error".to_owned() }
-                })?;
-            Ok(AccountTransaction::Declare(declare_tx))
-        }
-        RpcTransaction::DeployAccount(RpcDeployAccountTransaction::V3(tx)) => {
-            let deploy_account_tx = DeployAccountTransaction::V3(DeployAccountTransactionV3 {
-                resource_bounds: ValidResourceBounds::AllResources(tx.resource_bounds),
-                tip: tx.tip,
-                signature: tx.signature.clone(),
-                nonce: tx.nonce,
-                class_hash: tx.class_hash,
-                contract_address_salt: tx.contract_address_salt,
-                constructor_calldata: tx.constructor_calldata.clone(),
-                nonce_data_availability_mode: tx.nonce_data_availability_mode,
-                fee_data_availability_mode: tx.fee_data_availability_mode,
-                paymaster_data: tx.paymaster_data.clone(),
-            });
-            let contract_address = calculate_contract_address(
-                deploy_account_tx.contract_address_salt(),
-                deploy_account_tx.class_hash(),
-                &deploy_account_tx.constructor_calldata(),
-                ContractAddress::default(),
-            )
-            .map_err(|e| {
-                error!("Failed to calculate contract address: {}", e);
-                GatewaySpecError::UnexpectedError { data: "Internal server error".to_owned() }
-            })?;
-            let tx_hash = deploy_account_tx
-                .calculate_transaction_hash(chain_id, &deploy_account_tx.version())
-                .map_err(|e| {
-                    error!("Failed to calculate tx hash: {}", e);
-                    GatewaySpecError::UnexpectedError { data: "Internal server error".to_owned() }
-                })?;
-            let deploy_account_tx = BlockifierDeployAccountTransaction::new(
-                deploy_account_tx,
-                tx_hash,
-                contract_address,
-            );
-            Ok(AccountTransaction::DeployAccount(deploy_account_tx))
-        }
-        RpcTransaction::Invoke(RpcInvokeTransaction::V3(tx)) => {
-            let invoke_tx = starknet_api::transaction::InvokeTransaction::V3(InvokeTransactionV3 {
-                resource_bounds: ValidResourceBounds::AllResources(tx.resource_bounds),
-                tip: tx.tip,
-                signature: tx.signature.clone(),
-                nonce: tx.nonce,
-                sender_address: tx.sender_address,
-                calldata: tx.calldata.clone(),
-                nonce_data_availability_mode: tx.nonce_data_availability_mode,
-                fee_data_availability_mode: tx.fee_data_availability_mode,
-                paymaster_data: tx.paymaster_data.clone(),
-                account_deployment_data: tx.account_deployment_data.clone(),
-            });
-            let tx_hash = invoke_tx
-                .calculate_transaction_hash(chain_id, &invoke_tx.version())
-                .map_err(|e| {
-                    error!("Failed to calculate tx hash: {}", e);
-                    GatewaySpecError::UnexpectedError { data: "Internal server error".to_owned() }
-                })?;
-            let invoke_tx = BlockifierInvokeTransaction::new(invoke_tx, tx_hash);
-            Ok(AccountTransaction::Invoke(invoke_tx))
-        }
-    }
 }
 
 // TODO(yael 9/5/54): Should be implemented as part of InternalTransaction in starknet-api
