+++ crates/mempool_test_utils/src/starknet_api_test_utils.rs
@@ -133,6 +133,10 @@ pub fn test_resource_bounds_mapping() -> AllResourceBounds {
     )
 }
 
+pub fn test_valid_resource_bounds() -> ValidResourceBounds {
+    ValidResourceBounds::AllResources(test_resource_bounds_mapping())
+}
+
 /// Get the contract class used for testing.
 pub fn contract_class() -> ContractClass {
     env::set_current_dir(get_absolute_path(TEST_FILES_FOLDER)).expect("Couldn't set working dir.");
@@ -174,6 +178,14 @@ pub fn invoke_tx(cairo_version: CairoVersion) -> RpcTransaction {
     MultiAccountTransactionGenerator::new_for_account_contracts([default_account])
         .account_with_id(0)
         .generate_default_invoke()
+}
+
+pub fn executable_invoke_tx(cairo_version: CairoVersion) -> Transaction {
+    let default_account = FeatureContract::AccountWithoutValidations(cairo_version);
+
+    MultiAccountTransactionGenerator::new_for_account_contracts([default_account])
+        .account_with_id(0)
+        .generate_default_executable_invoke()
 }
 
 //  TODO(Yael 18/6/2024): Get a final decision from product whether to support Cairo0.
@@ -282,6 +294,17 @@ impl AccountTransactionGenerator {
             calldata: create_trivial_calldata(self.test_contract_address()),
         );
         rpc_invoke_tx(invoke_args)
+    }
+
+    pub fn generate_default_executable_invoke(&mut self) -> Transaction {
+        let invoke_args = starknet_api::invoke_tx_args!(
+            sender_address: self.sender_address(),
+            resource_bounds: test_valid_resource_bounds(),
+            nonce: self.next_nonce(),
+            calldata: create_trivial_calldata(self.test_contract_address()),
+        );
+
+        Transaction::Invoke(starknet_api::test_utils::invoke::executable_invoke_tx(invoke_args))
     }
 
     pub fn generate_default_deploy_account(&mut self) -> RpcTransaction {
